FEATURE: Begin DB implementation
There's still a lot of work and testing to do in order to set this up but we do have a trajectory. The remaining question that I have is primarily around JSON types in SQLite - is there a first-class JSON type in the flavor of SQLite that expo-sqlite is using? I only saw webSQL as a dependency, so my concerns are that I can't actually define a column as JSON and will need to fall back to  and then serialize.